requirements.txt
flask
flask-restful
flask-cors
python-dotenv
pylint
requests
gunicorn
gevent
bs4
twitterscraper
tqdm
flask-marshmallow
pymongo
validators
tldextract
dateparser
webargs
pytest
tavern